Add differential featureion information to a tbl using edgeR.

join_features() extracts and joins information for specific features

join_features(
  .data,
  features = NULL,
  all = FALSE,
  exclude_zeros = FALSE,
  shape = "long",
  ...
)

Arguments

.data
A tidy SingleCellExperiment object

features
A vector of feature identifiers to join

all
If TRUE return all

exclude_zeros
If TRUE exclude zero values

shape
Format of the returned table "long" or "wide"

...
Parameters to pass to join wide, i.e. assay name to extract feature abundance from

Value

A tbl containing the information.for the specified features

Details

This function extracts information for specified features and returns the information in either long or wide format.

Examples

tidySingleCellExperiment::pbmc_small %>%

    join_features(features=c("HLA-DRA", "LYZ"))

#> tidySingleCellExperiment says: This operation lead to duplicated cell names. A data frame is returned for independent data analysis.

#> # A tibble: 160 x 34
#>    cell  feature abundance_counts abundance_logco… orig.ident nCount_RNA
#>    <chr> <chr>              <dbl>            <dbl> <fct>           <dbl>
#>  1 ATGC… HLA-DRA                0             0    SeuratPro…         70
#>  2 ATGC… LYZ                    1             4.97 SeuratPro…         70
#>  3 CATG… HLA-DRA                1             4.78 SeuratPro…         85
#>  4 CATG… LYZ                    1             4.78 SeuratPro…         85
#>  5 GAAC… HLA-DRA                0             0    SeuratPro…         87
#>  6 GAAC… LYZ                    1             4.75 SeuratPro…         87
#>  7 TGAC… HLA-DRA                0             0    SeuratPro…        127
#>  8 TGAC… LYZ                    0             0    SeuratPro…        127
#>  9 AGTC… HLA-DRA                1             4.07 SeuratPro…        173
#> 10 AGTC… LYZ                    0             0    SeuratPro…        173
#> # … with 150 more rows, and 28 more variables: nFeature_RNA <int>,
#> #   RNA_snn_res.0.8 <fct>, letter.idents <fct>, groups <chr>,
#> #   RNA_snn_res.1 <fct>, file <chr>, ident <fct>, PC_1 <dbl>, PC_2 <dbl>,
#> #   PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, PC_6 <dbl>, PC_7 <dbl>, PC_8 <dbl>,
#> #   PC_9 <dbl>, PC_10 <dbl>, PC_11 <dbl>, PC_12 <dbl>, PC_13 <dbl>,
#> #   PC_14 <dbl>, PC_15 <dbl>, PC_16 <dbl>, PC_17 <dbl>, PC_18 <dbl>,
#> #   PC_19 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>